fix: remove dynamic from tuono-router
