+++ packages/tuono-router/src/index.ts
@@ -2,5 +2,4 @@ export { RouterProvider } from './components/RouterProvider'
 export { default as Link } from './components/Link'
 export { createRouter } from './router'
 export { createRoute, createRootRoute } from './route'
-export { dynamic, __tuono__internal__lazyLoadComponent } from './dynamic'
 export { useRouter } from './hooks/useRouter'
